Add new db type and some fake data sql file and removed idx from EventState and only call getEvent once and populate state and some other changes

+++ utils/utils.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 )
 
-func NumberToString(number int64) string {
+func NumberToString(number int) string {
 	return fmt.Sprintf("%d", number)
 
 }
